+++ changelog.md
@@ -13,6 +13,7 @@
     - `thumbnailOverlayBehavior: "never"` - replaced with `thumbnailOverlayEnabled`.
       - Setting the thumbnail overlay mode to `never` will now disable only the automatic thumbnail overlay, allowing for the toggle button to be used independently.
     - `autoScrollToActiveSongMode: "never"` - replaced with `autoScrollToActiveSongEnabled`.
+  - The welcome menu now allows all privacy-sensitive features to be toggled off before the script runs for the first time.
   - Made the version check feature more resilient to stop erroneous update notifications for weird versions like `5` from showing up.
 - **Plugin Changes:**
   - Added new features to the BYTM object. [Refer to the API docs for details.](https://github.com/Sv443/BetterYTM/blob/develop/contributing.md#global-functions-and-classes)
@@ -30,6 +31,8 @@
   - Added new events to the plugin interface:
     - `bytm:dataStoreSerializerLoaded` - Emitted after all memory-cached DataStore instances' data was lazy-loaded.
   - All DataStore instances that have in-memory cached data will now be lazy-loaded after feature initialization is done. Note: lazy-loading starts after `bytm:allReady`, so it's more aimed at lowering initial data access times.
+  - Features can now have tags associated with them, which is another way of filtering them, like when using the new internal function `configSetFeatsWithTags()`.
+  - Updated the dependencies CoreUtils to v3.8.0 and UserUtils to v11.0.0
 
 <div class="pr-link-cont">
   <a href="https://github.com/Sv443/BetterYTM/pull/TODO" rel="noopener noreferrer">Also see pull request #TODO</a>